% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack using brute force O(2^n)}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{dataframe containing v and w}

\item{W}{weight threshold}
}
\value{
value and objects that made up that value
}
\description{
knapsack greedy implementation
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500, parallel=FALSE)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500, parallel=TRUE)
}
