% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack2.R
\name{brute_force_knapsack2}
\alias{brute_force_knapsack2}
\title{Knapsack Algorithm - Brute Force Method 2}
\usage{
brute_force_knapsack2(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame with colnames 'v' & 'w'}

\item{W}{Knapsack Maximum Weight}
}
\value{
Return best Knapsack combination with maximum value
}
\description{
Brute Force Optional Approch to Solve Knapsack Problem
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 )
 brute_force_knapsack2(x = knapsack_objects[1:8,], W = 2000)

}
