% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{calc_comb}
\alias{calc_comb}
\title{calc_comb}
\usage{
calc_comb(vect, n, v, w)
}
\arguments{
\item{vect}{vector to apply the function for}

\item{n}{number of knapsack items}

\item{v}{values}

\item{w}{weights}
}
\value{
list with sum v, sum w and indeces for all combinations
}
\description{
Calculates value and weight of the combinations
}
