% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{find_indeces}
\alias{find_indeces}
\title{find_indeces}
\usage{
find_indeces(m, i, j, w, W)
}
\arguments{
\item{m}{matrix that for each index pair i and j specifies the maximum value
that can be attained with weight less than or equal to j using items
up to first i items. It is calulated in function knapsack_dynamic}

\item{i}{number of elements plus 1}

\item{j}{maximum weight plus 1}

\item{w}{vector with weight values}

\item{W}{maximum weight}
}
\value{
vector with indeces

examples see function knapsack_dynamic
}
\description{
Help function to function knapsack_dynamic, finds the indeces of the elements
that gives the solution with highest value given the maximum weight W
restriction
}
