% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knapsack using greedy heuristic to reduce complexity}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{dataframe containing v and w}

\item{W}{weight threshold}
}
\value{
value and objects that made up that value
}
\description{
implementation of knapsack greedy algorithm
}
\examples{
n<-10000
knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),v=runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
