% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{knapsack_brute_force}
\usage{
knapsack_brute_force(x, W, parallel = FALSE)
}
\arguments{
\item{x}{x is a dataframe containing 2 columns, positive weights and values of the objects.}

\item{W}{The maximum weight of the knapsack.}

\item{parallel}{TRUE or FALSE, decides if parallel computation should be used. Might not be faster on all computers.}
}
\value{
The value of the best knapsack and the elements included in it.
}
\description{
Solution to the knapsack problem that runs in O(2^n) time.
}
\examples{

data("knapsack_objects")
knapsack_brute_force(x = knapsack_objects[1:8,], W = 2000)


}
