% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic_knapsack.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack using Dynamic Programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{dataframe containing v and w}

\item{W}{weight threshold}
}
\value{
value and objects that made up that value
}
\description{
implementation of knapsack ysing dynamic programing algorithm
}
\examples{
n<-10000
knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),v=runif(n = n, 0, 10000))
knapsack_dynamic(x = knapsack_objects[1:800,], W = 3500)
}
