% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsackdynamicrcpp.r
\name{knapsack_dynamicrcpp}
\alias{knapsack_dynamicrcpp}
\title{Knapsack Problem using the dynamic approach using Rcpp}
\usage{
knapsack_dynamicrcpp(x, W)
}
\arguments{
\item{x}{value}

\item{W}{weight}
}
\value{
A list giving total no of objects and the maximum weight it can hold
\itemize{
\item x -knapsack_objects
\item W - total weight it can hold
}
}
\description{
Knapsack Problem using the dynamic approach using Rcpp
}
\examples{
knapsack_dynamicrcpp(x = knapsack_objects[1:8,], W = 2000)
}
