% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_greedy.R
\name{knapsack_greedy}
\alias{knapsack_greedy}
\title{knapsack_greedy}
\source{
https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm
}
\usage{
knapsack_greedy(x, W)
}
\arguments{
\item{x}{x is a dataframe containing 2 columns, positive weights and values of the objects.}

\item{W}{The maximum weight of the knapsack.}
}
\value{
The value of the best knapsack and the elements included in it.
}
\description{
An approximation to the knapsack problem that runs in O(n log(n)) time.
}
\examples{

suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE), v = runif(n = n, 0, 10000))
knapsack_greedy(x = knapsack_objects[1:1200,], W = 2000)


}
