% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{knapsack_objects}
\format{
A data frame with columns v and w.
\describe{
Each row describes an object with a value (v) and weight (w).
}
}
\source{
See \code{inst/scripts/create-data.R} for data preparation.
}
\usage{
data(knapsack_objects)
}
\description{
This contains an example dataset that is used to create and test the lab.
}
\keyword{datasets}
