% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_brute.R
\name{old_knapsack_brute_force}
\alias{old_knapsack_brute_force}
\title{knapsack_brute_force}
\usage{
old_knapsack_brute_force(x, W)
}
\arguments{
\item{x}{x is a dataframe containing 2 columns, positive weights and values of the objects.}

\item{W}{The maximum weight of the knapsack.}
}
\value{
The value of the best knapsack and the elements included in it.
}
\description{
Solution to the knapsack problem that runs in O(2^n) time.
}
\examples{

suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE), v = runif(n = n, 0, 10000))
knapsack_dynamic(x = knapsack_objects[1:8,], W = 2000)


}
