\name{Team05Lab06-package}
\alias{Team05Lab06-package}
\alias{Team05Lab06}
\docType{package}
\title{
\packageTitle{Team05Lab06}
}
\description{
\packageDescription{Team05Lab06}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Team05Lab06}
\packageIndices{Team05Lab06}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Team05Lab06}

Maintainer: \packageMaintainer{Team05Lab06}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
