% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, type = "forloop", filter_items = FALSE, ncores = 2)
}
\arguments{
\item{x}{data frame with two variables v and w}

\item{W}{knapsack size}

\item{type}{type of calculation}

\item{filter_items}{logical indicating}

\item{ncores}{number of cores to use in parallel computing (only applies for
types = "parallel1a", "parallel1b" and "parallel2"). Note that
default is set to two because of CRAN limit.}
}
\value{
list with value and elements
}
\description{
Implements several methods to run the brute force algorithm for the knapsack
problem
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 )
x <- knapsack_objects[1:8,]
brute_force_knapsack(x = x, W = 3500)
brute_force_knapsack(x = x, W = 3500, type="matrix", filter_items=TRUE)
}
