% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force solution for the knapsack problem}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data structure of weights and values}

\item{W}{max weight}

\item{parallel}{TRue if with parallel, False (default if not)}
}
\value{
most valued combination of elements which fits backpack and the maximum value achieved
}
\description{
Brute force solution for the knapsack problem
}
