% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruteforce.r
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack problem using the bruteforce approach}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{value}

\item{W}{weight}

\item{parallel}{for parallel computation}
}
\value{
A list giving total no of objects and the maximum weight it can hold
\itemize{
\item x -knapsack_objects
\item W - total weight it can hold
}
}
\description{
Knapsack problem using the bruteforce approach
}
