% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack Algorithm - Brute Force}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame with colnames 'v' & 'w'}

\item{W}{Knapsack Maximum Weight}

\item{parallel}{Bool value to pass to active parallel}
}
\value{
Return best Knapsack combination with maximum value
}
\description{
Brute Force Approch to Solve Knapsack Problem
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 )
 brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)

}
