% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knapsack greedy approximation}
\usage{
greedy_knapsack(x, W, type = "normal")
}
\arguments{
\item{x}{is the data frame with the items value and weigth knapsack objects}

\item{W}{is the weight limit of the knap sack}

\item{type}{it runs the normal or plus(s+1) type or greedy}
}
\value{
List with values and elements
}
\description{
Knapsack greedy approximation
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 100000
knapsack_objects <-
data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 )
x <- knapsack_objects[1:8,]
greedy_knapsack(x = x, W = 3500)
greedy_knapsack(x = x, W = 3500, type = "plus")
}
