% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Return maximum value of items and their fractional amounts.}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data structure of weights and values}

\item{W}{max weight}
}
\value{
most valued combination of elements which fits backpack
}
\description{
Return maximum value of items and their fractional amounts.
}
