% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy solution for the knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data structure of weights and values}

\item{W}{max weight}
}
\value{
most valued combination of elements which fits backpack and the maximum value achieved
}
\description{
Greedy solution for the knapsack problem
}
