% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.r
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{value}

\item{W}{weight}
}
\value{
A list giving total no of objects and the maximum weight it can hold
\itemize{
\item x -knapsack_objects
\item W - total weight it can hold
}
}
\description{
Knapsack Problem using the greedy approach
}
\references{
'https://en.wikipedia.org/wiki/Knapsack_problem'
order data decrasely by value(i)/weight(i)
Problem: value(i)/weight(i) is equal but value(i) is not equal
should order decreasely by value(i) too.
print(new_data)
}
