% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyknapsack.r
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knapsack Problem using the greedy approach}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{value}

\item{W}{weight}
}
\value{
A list giving total no of objects and the maximum weight it can hold
\itemize{
\item x -knapsack_objects
\item W - total weight it can hold
}
}
\description{
Knapsack Problem using the greedy approach
}
\examples{
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
order data decrasely by value(i)/weight(i)
Problem: value(i)/weight(i) is equal but value(i) is not equal
should order decreasely by value(i) too.
print(new_data)
}
