% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data structure of weights and values}

\item{W}{max weight}
}
\value{
most valued combination of elements which fits backpack
}
\description{
knapsack_dynamic
}
