% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.r
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack problem using the dynamic approach}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{value}

\item{W}{weight}
}
\value{
A list giving total no of objects and the maximum weight it can hold
\itemize{
\item x -knapsack_objects
\item W - total weight it can hold
}
}
\description{
Knapsack problem using the dynamic approach
}
\examples{
\\donotrun{

knapsack_dynamic (x = knapsack_objects[1:8,], W = 2000)
}
}
