% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsackdynamicrcpp.r
\name{knapsack_dynamicrcpp}
\alias{knapsack_dynamicrcpp}
\title{Knapsack Problem using the dynamic approach using Rcpp
knapsack_dynamicrcpp(x = knapsack_objects[1:8,], W = 2000)}
\usage{
knapsack_dynamicrcpp(x, W)
}
\description{
Knapsack Problem using the dynamic approach using Rcpp
knapsack_dynamicrcpp(x = knapsack_objects[1:8,], W = 2000)
}
